// server/src/models/DisciplineType.ts
/**
 * Types of allied health disciplines supported by the system
 */
export enum DisciplineType {
  PHYSIOTHERAPY = 'PHYSIOTHERAPY',
  OCCUPATIONAL_THERAPY = 'OCCUPATIONAL_THERAPY',
  SPEECH_PATHOLOGY = 'SPEECH_PATHOLOGY',
  DIETETICS = 'DIETETICS',
  PODIATRY = 'PODIATRY',
  CHIROPRACTIC = 'CHIROPRACTIC',
  OSTEOPATHY = 'OSTEOPATHY',
  PSYCHOLOGY = 'PSYCHOLOGY',
  EXERCISE_PHYSIOLOGY = 'EXERCISE_PHYSIOLOGY',
  OTHER = 'OTHER'
} 